# Conflicts:
#	segmentstore/server/src/main/java/io/pravega/segmentstore/server/store/ServiceConfig.java
#	segmentstore/storage/src/test/java/io/pravega/segmentstore/storage/chunklayer/SystemJournalTests.java

+++ segmentstore/server/src/test/java/io/pravega/segmentstore/server/writer/StorageWriterTests.java
@@ -1021,7 +1021,7 @@ public class StorageWriterTests extends ThreadPooledTestSuite {
 
             // Add the operation to the log.
             StreamSegmentMapOperation mapOp = new StreamSegmentMapOperation(context.storage.getStreamSegmentInfo(name, TIMEOUT).join());
-            mapOp.setStreamSegmentId((long) i);
+            mapOp.setStreamSegmentId(i);
             context.dataSource.add(mapOp);
         }
